pyproject.toml
[build-system]
requires = ["setuptools >= 61"]
build-backend = "setuptools.build_meta"

[project]
name = "bia_bmz_integrator"
version = "0.1.0"
description = "Integration of BioImage Archive and BioImage Model Zoo."
readme = "README.md"
license = { text = "" }
authors = [
    {name = "Teresa Zulueta Coarasa"},
    {name = "Craig Russell"}, 
    {name = "Christopher Rookyard"},
]
requires-python = ">=3.10"
dependencies = [
    "bioimageio-core", 
    "bioimageio-spec", 
    "click", 
    "dask", 
    "matplotlib", 
    "monai", 
    "numpy", 
    "opencv-python", 
    "pandas", 
    "requests",
    "torch", 
    "torchvision", 
    "zarr",  
]

[tool.setuptools]
packages = ["bia_bmz_integrator", "bia_bmz_integrator.data_models", "bia_bmz_integrator.process", "bia_bmz_integrator.scripts"]
include-package-data = true

[project.scripts]
bia_bmz_benchmark = "bia_bmz_integrator.scripts.run_benchmark_bmz_on_bia_data:cli"
amalgamate_jsons = "bia_bmz_integrator.scripts.amalgamate_jsons:amalgamate_jsons"
